requirements.txt
Django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
PyJWT==1.6.4
pytz==2018.5
